refactor(popup): group the recording controls under src/popup/recording/
Moves recording.ts, recording-io.ts and action-recording.ts into
src/popup/recording/ — the popup's two recording surfaces (tab/screen capture
and action-workflow capture) plus the chrome.runtime/storage/mic-permission
IO layer they sit on.

Evidence: recording-io.ts has exactly one importer, recording.ts, and was
extracted from it; the pair is already treated as one unit by
tests/extension/recording-log-branding.test.js, which asserts the rule-21
shared-label invariant across both files at once. action-recording.ts joins
them because it is the third module that writes user-facing recording labels
and the only other popup module that reaches lib/brand.ts for the log prefix —
a wording change to "recording" has to touch all three together and nothing
else in src/popup.

Filename similarity was not the deciding factor here and would in fact have
been misleading: "recording.ts" and "action-recording.ts" are different
features (tab-recording/playback-engine vs flow-recording, per their Docs
headers) that share a label vocabulary, while "recording-io.ts" reads like a
generic IO utility and is actually specific to one caller.

Non-obvious path references updated:
  - tests/extension/action-recording-reconcile.test.js (compiled
    extension/popup/... dynamic import)
  - tests/extension/recording-log-branding.test.js (src/... string literals)
  - flow-recording and tab-recording feature indexes; docs/core/
    bug-class-audit.md finding F5, which names the file by path

12 stale compiled artifacts under extension/popup/ were git rm'd; the rebuilt
output lands in extension/popup/recording/.
