+++ extension/popup.js
@@ -11,9 +11,9 @@ import { updateConnectionStatus } from './popup/shell/status-display.js';
 import { renderUpdateAvailableBanner } from './popup/update-button.js';
 import { DEFAULT_SERVER_URL } from './lib/constants.js';
 import { buildDaemonHeaders } from './lib/daemon-http.js';
-import { setupRecordingUI } from './popup/recording.js';
+import { setupRecordingUI } from './popup/recording/recording.js';
 import { setupDrawModeButton } from './popup/draw-mode.js';
-import { setupActionRecordingUI } from './popup/action-recording.js';
+import { setupActionRecordingUI } from './popup/recording/action-recording.js';
 import { FEATURE_TOGGLES as TOGGLE_DEFS, applyFeatureToggles } from './popup/feature-toggles.js';
 import { initTrackPageButton } from './popup/tab-tracking.js';
 import { applyAiWebPilotToggle } from './popup/ai-web-pilot.js';
