+++ extension/popup/recording/action-recording.d.ts
