+++ extension/popup/recording/action-recording.js
@@ -3,9 +3,9 @@
  * Why: Separates event recording controls from screen recording, keeping each feature self-contained.
  * Docs: docs/features/feature/flow-recording/index.md
  */
-import { DEFAULT_SERVER_URL, StorageKey } from '../lib/constants.js';
-import { postDaemonJSON } from '../lib/daemon-http.js';
-import { getLocal, setLocal, removeLocal, persist } from '../lib/storage-utils.js';
+import { DEFAULT_SERVER_URL, StorageKey } from '../../lib/constants.js';
+import { postDaemonJSON } from '../../lib/daemon-http.js';
+import { getLocal, setLocal, removeLocal, persist } from '../../lib/storage-utils.js';
 const START_LABEL = 'Record action workflow';
 const STOP_LABEL = 'Stop recording';
 function showRecording(els, state) {
