+++ extension/popup/recording/recording.d.ts
