+++ src/popup/recording/recording.ts
@@ -10,9 +10,9 @@
  * Manages recording controls, timer display, and mic permission flow.
  */
 
-import { StorageKey } from '../lib/constants.js'
-import { KABOOM_RECORDING_LOG_PREFIX } from '../lib/brand.js'
-import { getLocal, removeLocal, onStorageChanged, persist } from '../lib/storage-utils.js'
+import { StorageKey } from '../../lib/constants.js'
+import { KABOOM_RECORDING_LOG_PREFIX } from '../../lib/brand.js'
+import { getLocal, removeLocal, onStorageChanged, persist } from '../../lib/storage-utils.js'
 import {
   sendRecordingGestureDecision,
   handleStartClick,
